Fix Windows warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@61957 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 81053f1c00aed1cfb729d8c169639b07b8b349ee

+++ LHCb/Sim/SimComponents/src/CheckMCEventTool.h
@@ -1,4 +1,4 @@
-// $Id: CheckMCEventTool.h,v 1.2 2008-10-28 15:34:08 cattanem Exp $
+// $Id: CheckMCEventTool.h,v 1.3 2010-05-18 22:04:01 cattanem Exp $
 #ifndef CHECKMCEVENTTOOL_H 
 #define CHECKMCEVENTTOOL_H 1
 
@@ -31,7 +31,7 @@ public:
 protected:
 
 private:
-  int m_event;         ///< last event number
+  longlong m_event;    ///< last event number
   StatusCode m_status; ///< Status of last check
 };
 #endif // CHECKMCEVENTTOOL_H
